+++ tests/commands/test_convert_to_oc4ids.py
@@ -6,3 +6,9 @@ def test_command(monkeypatch):
     assert_streaming(monkeypatch, main, ['convert-to-oc4ids', '--project-id', '1'],
                      ['release_1.1.json'],
                      ['oc4ids-project_minimal.json'])
+
+
+def test_command_package(monkeypatch):
+    assert_streaming(monkeypatch, main, ['convert-to-oc4ids', '--project-id', '1', '--package'],
+                     ['release_1.1.json'],
+                     ['oc4ids-project-package_minimal.json'])