+++ tests/commands/test_package_releases.py
@@ -4,7 +4,7 @@ from tests import assert_streaming
 
 def test_command(monkeypatch):
     assert_streaming(monkeypatch, main, ['package-releases', '--uri', 'http://example.com', '--published-date',
-                                         '9999-01-01T00:00:00Z', '--publisher-name', ''],
+                                         '9999-01-01T00:00:00Z', '--version', '1.2', '--publisher-name', ''],
                      ['release_minimal-1.json', 'release_minimal-2.json'],
                      ['release-package_minimal-1-2.json'])
 
@@ -17,7 +17,7 @@ def test_command_size(monkeypatch):
 
 def test_command_extensions(monkeypatch):
     assert_streaming(monkeypatch, main, ['package-releases', '--uri', 'http://example.com', '--published-date',
-                                         '9999-01-01T00:00:00Z', '--publisher-name', '',
+                                         '9999-01-01T00:00:00Z', '--version', '1.2', '--publisher-name', '',
                                          'http://example.com/a/extension.json', 'http://example.com/b/extension.json'],
                      ['release_minimal-1.json', 'release_minimal-2.json'],
                      ['release-package_minimal-1-2-extensions.json'])